refactor: move shared runtime-metrics dispatch out of jvm.go
